created abstract product, presetup project with bit more classes
