+++ ExamPub/src/main/java/com/urban/exampub/controllers/UserController.java
@@ -1,0 +1,8 @@
+package com.urban.exampub.controllers;
+
+import org.springframework.stereotype.Controller;
+
+@Controller
+public class UserController {
+
+}
