+++ ExamPub/src/main/java/com/urban/exampub/models/Drink.java
@@ -3,39 +3,21 @@ package com.urban.exampub.models;
 import jakarta.persistence.*;
 
 @Entity
-public class Drink {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String productName;
-    private double price;
+public class Drink extends Product{
+
     private boolean isForAdult;
 
     public Drink() {
     }
 
     public Drink(String productName, double price, boolean isForAdult) {
-        this.productName = productName;
-        this.price = price;
+        super(productName, price);
         this.isForAdult = isForAdult;
     }
-
-    public String getProductName() {
-        return productName;
+    @Id
+    public Long getId() {
+        return super.getId();
     }
-
-    public void setProductName(String productName) {
-        this.productName = productName;
-    }
-
-    public double getPrice() {
-        return price;
-    }
-
-    public void setPrice(double price) {
-        this.price = price;
-    }
-
     public boolean isForAdult() {
         return isForAdult;
     }
@@ -44,11 +26,4 @@ public class Drink {
         isForAdult = forAdult;
     }
 
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public Long getId() {
-        return id;
-    }
 }
