+++ ExamPub/src/main/java/com/urban/exampub/repositories/DrinkRepository.java
@@ -1,0 +1,8 @@
+package com.urban.exampub.repositories;
+
+import com.urban.exampub.models.Drink;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface DrinkRepository extends JpaRepository<Drink, Long> {}
