+++ ExamPub/src/main/java/com/urban/exampub/repositories/OrderRepository.java
@@ -1,0 +1,10 @@
+package com.urban.exampub.repositories;
+
+import com.urban.exampub.models.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+interface OrderRepository extends JpaRepository<Order, Long> {
+
+}
