+++ ExamPub/src/main/java/com/urban/exampub/repositories/UserRepository.java
@@ -1,0 +1,8 @@
+package com.urban.exampub.repositories;
+
+import com.urban.exampub.models.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<User, Long> {}
