+++ ExamPub/src/main/java/com/urban/exampub/services/OrderService.java
@@ -1,0 +1,6 @@
+package com.urban.exampub.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+interface OrderService {}
