+++ ExamPub/src/main/java/com/urban/exampub/services/UserServiceImpl.java
@@ -1,0 +1,6 @@
+package com.urban.exampub.services;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserServiceImpl implements UserService {}
